FINALLY FIXED THE GRID BUG!
Not ideal but works flawlessly. First I removed the last of child padding. Then I added the gutter padding to both the left and right and divided it by 2.  This balanced out all the spacing in the grid to preform properly minus the added padding to the sides of each row. So i added a negative margin to the row that is equal to a total of the gutter padding divided by 2.